+++ packages/in_app_purchase/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.1+1
 
-* Android: Do not register if Activity is null.
+* Android: Require Activity only to use launchBillingFlow method.
 
 ## 0.2.1